Renaming linux-specifix file with tests
Files are globbed based on their name. Therefore, we have to rename
them so tests for Linux are not executed on a Mac.
