+++ Library/Homebrew/test/os/dependency_collector_spec.rb
@@ -1,1 +1,1 @@
-require "test/os/linux/dependency_collector_spec" if OS.linux?
+require "test/os/linux/dependency_collector" if OS.linux?
